+++ app/routers/__init__.py
@@ -1,20 +1,24 @@
 """FastAPI Routers"""
 from app.routers import (
-    fields,
-    objects,
-    object_fields,
-    records,
-    relationships,
-    relationship_records,
     applications,
+    auth,
+    dashboard,
+    fields,
+    object_fields,
+    objects,
+    records,
+    relationship_records,
+    relationships,
 )
 
 __all__ = [
+    "applications",
+    "auth",
+    "dashboard",
     "fields",
     "objects",
     "object_fields",
     "records",
     "relationships",
     "relationship_records",
-    "applications",
 ]